Factored only those awarded plus rivers with potential.
Our goal here is to double check first which are awarded and then get company financial status which would determine who are good and bad. Our GOD Willing moment would be acquisition of a certain company which has a good number of Dams (hopefully in a certain river) which would make Repower venture in a different location.